requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
distlib==0.3.9
fastapi>=0.101.0
filelock==3.17.0
idna==3.10
packaging==24.2
pipenv==2024.4.1
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
virtualenv==20.29.1

uvicorn>=0.23.2
sqlalchemy>=2.0.20
pyjwt>=2.8.0
python-multipart>=0.0.6
bcrypt>=4.0.1
aiofiles>=23.1.0
jinja2>=3.1.2